split up config module
